Add method for converting integer values to arraylist
Created a new method that converts integer values to arraylist of digits
to solve issues with starting input to a new equation
with the solution from the previous equation. Previous the solution was
stored in an integer variable called 'left_operand' however this caused
issues if the user pressed backspace to delete a digit immidately after
a solution was returned. Now that the solution is stored in
'input_digits_left_operand' the conflict is resolved.