Add check for decimal.
Added boolean statement to check if the user has already set a decimal
place. This prevents the user from adding mulitple -1 values, which
represent a decimal point, to the arrayList of input digits. Mulitple
decimal values broke other functionality such as backspace delete.
Also, added check to the backspace delete function to set the boolean
back to false if the user deletes the decimal point.